Remove validate_spec from public API
